// react-forms/src/types/sign-up-form-fields-values.type.ts
export default interface ISignUpFormFieldsValues {
  fullName: string;
  email: string;
  password: string;
  dateOfBirth: string;
  gender: string;
  paymentMethod: string;
  cardNumber: string;
  cardCvv: string;
  cardExpireMonth: string;
  cardExpireYear: string;
  termsAndConditions: boolean;
  marketingInfo: boolean;
};